add the basic use of dataset and llm...

+++ backend/ranking/models.py
@@ -7,5 +7,5 @@ from testing import models as testing
 class Credit(models.Model):
     dataset = models.ForeignKey(to=dataset.Dataset, on_delete=models.CASCADE)
     LLM = models.ForeignKey(to=testing.LLMs, on_delete=models.CASCADE)
-    credit = models.FloatField(null=True)
+    credit = models.IntegerField(null=True)
     add_time = models.DateTimeField(auto_now_add=True)
